Refactor idpgis load testing tool
Break out the converters from the summarizer.
PEP8 work.

Add setup.cfg to idpgis load testing tool.

Fix final error rate calculation, there's no need to divide by 1000.

Fix overal error rate calculation in load testing tool.  The existing
calculation is brittle because it depends upon the status code returned
by arcgis, which is not to be trusted.  A better way is to look at the
return datatype, which is "bin" if we get back an image (success) or
"text" if we get back an error message (failure).