Add update-iso script for metadata generation
The purpose of this script is to update an existing ISO 19115 record
(XML) from both the YAML configuration file and from the REST page
metadata.

Remove gmd:maintenanceNote,  this was kind of goofy.  More confusing
than anything else and would just draw unwanted attention.

Use pathlib instead of os module for file management.

Remove noaa geoplatform support.  This is difficult to put into a YAML
config file...  And it only holds for a few records anyway.  Best to
handle this manually.